[front] chore: split Poke conversation page
